% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{dualModelFitness}
\alias{dualModelFitness}
\title{Fitness dual model function}
\usage{
dualModelFitness(vars, x, y, p = 0.8, cntr = trainControl(method =
  "LOOCV"), method = "rf")
}
\arguments{
\item{vars}{models variables}

\item{x}{A data.frame where samples are in rows
and features are in columns.}

\item{y}{A data.frame containing validation value}

\item{p}{A numeric value corresponding to randomly sample
a percentage of rows within a data frame}

\item{cntr}{Control the computational nuances of the train function
 (for more detail see \code{\link{trainmod}}).
By default, is used the cross-validation method \code{"LOOCV"}.}

\item{method}{A string specifying which classification and/or
regression model to use (for details see \code{\link{trainmod}}.
By default, is used the method \code{"rf"}.
Possible values are found using names (\code{\link[caret]{getModelInfo}}).}
}
\value{
numeric string describing putative variables ''fitness''.
}
\description{
The fitness dual model function has as input moldels variable
representing a potential solution and returns a numerical value
describing its ''fitness''. This function could be used in a genetic algorithm (
for more details see \code{\link[GA]{ga}})
}
