% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainmodel.R
\name{error}
\alias{error}
\title{Measurement error models}
\usage{
error(object, obs = NULL, newdata = NULL, type = c("se", "mae",
  "rmse", "rsq"), ...)
}
\arguments{
\item{object}{An object of \code{\link{Modelling-class}}.
This method use a numeric vector of observated outcomes.}

\item{obs}{a numeric vector of observated outcomes.}

\item{newdata}{An optional data-frame in which to look for variables
with which to predict. If omitted, the fitted values are used.}

\item{type}{A string of character that defines a . Current possibilities
are "se", "mae, "rmse" and "rsq". Type must be specified}

\item{...}{Additional parameters to be passed the the S4 method}
}
\description{
These functions are appropriate for cases where the model outcome is a number.
Given an object of \code{\link{Modelling-class}}, standard error (se),
mean absolute error (mae), Root Mean Squared Error (rmse), or R-squared (rsq)
is calculated.
}
