% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{gadualModel}
\alias{gadualModel}
\title{Genetic Algorithm}
\usage{
gadualModel(data_x, data_y, train, type = "binary",
  fitness = dualModelFitness, nBits = ncol(data_x), popSize = 20,
  elitism = 2, run = 10, maxiter = 500, monitor = plot,
  parallel = T, names = colnames(data_x), seed = 11, ...)
}
\arguments{
\item{data_x}{a data.frame}

\item{data_y}{numeric string}

\item{train}{a df.}

\item{type}{The type of genetic algorithm depending
on the nature of decision variables. They cpul be "binary" (default), "real-valued"
and "permutation". For more details see \code{\link[GA]{ga}}}

\item{popSize}{The population size.}

\item{elitism}{The number of best fitness individuals to survive at each generation}

\item{...}{Additional arguments to be passed to the fitness function}
}
\value{

}
\description{
for more details see \code{\link[GA]{ga}}
}
