% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{ImpMatrix}
\alias{ImpMatrix}
\title{Computing variable importance for each}
\usage{
ImpMatrix(train, validation)
}
\arguments{
\item{validation}{}
}
\value{
importance matrix
}
\description{
Extracting variable importance
and building a matrix. The importance of
predictors is reported in each column.
The name of the column is the name of
the predicted variables whereas the values
in each row correspond to the importance of
the variable. If I want to know the importance
of the 5th variable in predicting the
1st i have to extract the value in the 1st column
at the 5th row.
}
