% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainmodel.R
\docType{class}
\name{Modelling-class}
\alias{Modelling-class}
\alias{Modelling}
\title{The S4 class for Classification and Regression information.}
\description{
This function generates a valid references to a class.
}
\section{Slots}{

\describe{
\item{\code{varobs}}{data.frame. A data.frame containing observated values.}

\item{\code{yobs}}{numeric. A numeric vector of observated values used for training
data.frame.}

\item{\code{classification}}{train. Classification model.}

\item{\code{regression}}{train. Regression models.}
}}

