% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainmodel.R
\docType{methods}
\name{predict,Modelling-method}
\alias{predict,Modelling-method}
\title{Classifciation and/or Regression Models Prediction}
\usage{
\S4method{predict}{Modelling}(object, newdata = NULL, ...)
}
\arguments{
\item{Modelling}{For predict, an object of class \code{\link{Modelling-class}}.}
}
\value{
A numeric vector of predicted outcomes.
}
\description{
These functions can be used for a single train object or
to loop through a number of \code{\link{Modelling-class}} objects to calculate
the training and test data predictions and class probabilities.
The newdata (an optional set of data to predict on - see
{\code{\link[stats]{predict}}}) is NULL.
If not NULL, newdata training data are used.
For more detail see {\code{\link[stats]{predict}}}
}
