% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainmodel.R
\name{trainmod}
\alias{trainmod}
\title{Fit and Tune Predective Models and Select the "best" one}
\usage{
trainmod(x, y, cntr = caret::trainControl(method = "LOOCV"),
  method = "rf", ...)
}
\arguments{
\item{x}{A data.frame where samples are in rows
and features are in columns.}

\item{y}{A numeric vector containing the observated outcome for each sample.}

\item{cntr}{Control the computational nuances of the train function
 (for more detail see {\code{\link[caret]{trainControl}}}).
By default, is used the cross-validation method \code{"LOOCV"}.}

\item{method}{A string specifying which classification and/or
regression model to use (for details see \code{\link{caret}}.
By default, is used the method \code{"rf"}.
Possible values are found using names (\code{\link[caret]{getModelInfo}}).}

\item{...}{Arguments passed to the classification and regression
algorithm (such as \code{\link{randomForest}})}
}
\value{
a Modelling S4 Class is returned;
for details see \code{\link{Modelling-class}}
}
\description{
This function sets up a grid of tuning parametres for
classification and regression model. This function return an object of
\code{\link{Modelling-class}}.
}
\examples{

}
