% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableImportance.R
\name{variableImp}
\alias{variableImp}
\title{Variable importance through shuffle method}
\usage{
variableImp(object, obs = NULL, newdata = NULL, shuffles = 500,
  res.fun = mean)
}
\arguments{
\item{object}{An object of \code{\link{Modelling-class}}.}

\item{obs}{optional vector of real values}

\item{newdata}{an optional data.frame containing all variables
used in the object. If omitted, the fitted values are used}

\item{shuffles}{number of shuffles per variable}

\item{res.fun}{function for returning a single value per
variable after shuffling}
}
\value{
a vector containing the importance of each variable
 normalized by the RMSE of the full-model
}
\description{
The importance of each variable is computed
by randomizing its values and predicting.
A reference RMSE is computed using the
full-model whereas the RMSE of the shuffle
model is used for estimating variable importance.
}
